+++ R/eyes.R
@@ -1,32 +1,40 @@
 #' Count patients and eyes
 #' @name eyes
-#' @description Counts number of patients and right and left eyes. Columns are
+#' @description Counts number of subjects and right and left eyes. Columns are
 #'     guessed.
 #' @param x required. (data frame)
-#' @param id Patient identifying column, passed as (quoted) character
-#' @param eye Eye identifying column, passed as (quoted) character
+#' @param id_col Subject identifying column, passed as (quoted) character
+#'   Can also be abbreviated to "id" as per partial matching
+#' @param eye_col Eye identifying column, passed as (quoted) character.
+#'   Can also be abbreviated to "eye" as per partial matching
 #' @param dropunknown introduces NA for values not recognized by [recodeye]
-#' @section Guessing:
-#' **id** and **eye** arguments overrule the name guessing for
-#' the respective columns (here, cases need to match)
+#' @param details if TRUE, will add information about which and how many
+#'   subjects have only one eye or both included, and provide a list of
+#'   subject IDs for each
+#' @section Column guessing:
+#' **id_col** and **eye_col** arguments overrule the name guessing for
+#' the respective columns (here, cases need to match). Both arguments can be
+#' abbreviated (*id* or *eye*) as per partial argument name matching.
 #'
 #' For any below, **cases are always ignored** (column names can be in upper or
 #'    lower case, as you please)
 #'
 #' **patient ID columns**:
 #' - First, `eyes` is looking for names that contain both strings
-#' "pat" and "id" (the order doesn't matter)
+#' "pat" and "id" (the order doesn't matter) -
+#'   you can change this with [set_eye_strings]
 #' - Next, it will look for columns that are plainly called "ID"
 #' - Last, it will search for all names that contain either "pat"
 #' or "id"
 #'
-#' **eye variable column**:
+#' **eye column**:
 #' - `eyes` primarily looks for columns called either "eye" or "eyes",
+#' (you can change this with [set_eye_strings])
 #' and if they are not present, columns containing string "eye"
 #' (e.g., EyeName will be recognized)
 #'
 #' @section Eye coding:
-#' The following codes are recognized:
+#' The following codes are recognized: (change this with [set_eye_strings])
 #'
 #' - integer coding 0:1 and 1:2, right eye being the lower number.
 #' - right eyes: c("r", "re", "od", "right") and
@@ -36,65 +44,62 @@
 #' If your eye column contains other values, they will be dropped to
 #' NA (dropunknown) or kept (and then only patients will be counted, because
 #' coding remains unclear). Recommend then to recode with [recodeye]
-#' @section eyestr:
-#'  eyestr is a wrapper around eyes. It parses the result into meaningful text,
-#'  which can be #'  readily pasted into reports with [eyes_to_string] under the hood.
-#'  Arguments to `eyes_to_string` are passed via **...**:
-#'  - **small_num** If TRUE (default): numbers <= 12 as words
-#'  - **UK** TRUE: UK style (English) or FALSE (default):
-#'  US style (American).
-#' @return `eyes`: Named integer vector with count of patients and eyes
+#' @return List (of class "eyes" with count of patients and eyes
+#'   if "details = TRUE", an list of class "eyes_details" will be returned
 #' @family eye core functions
 #' @importFrom purrr quietly
 #' @examples
 #' library(eyedata)
 #' eyes(amd2)
 #'
+#' ## If you code your eyes with different strings,
+#' ## e.g., because you are using a different language,
+#' ## you can change this either with `set_eye_strings`
+#' set_eye_strings(right = c("droit", "od"), left = c("gauche", "og"))
+#'
+#' ## restore defaults with
+#' set_eye_strings()
 #' @export
 #'
-eyes <- function(x, id = NULL, eye = NULL, dropunknown = TRUE) {
+eyes <- function(x, id_col = NULL, eye_col = NULL, dropunknown = TRUE, details = FALSE) {
   if (!inherits(x, "data.frame")) {
     stop("x must be a data frame", call. = FALSE)
   }
   if (nrow(x) < 1) {
-    return("0 eyes of 0 patients")
+    return("0 eyes of 0 subject")
   }
-  if (is.null(id)) {
-    id <- getElem_id(x)
+  if (is.null(id_col)) {
+    id_col <- getElem_id(names(x))
   }
 
-  if (length(id) != 1) {
-    warning("Did not find the ID column - use argument \"id\"",
+  if (length(id_col) != 1) {
+    warning("Did not find the ID column - use argument \"id_col\"",
       call. = FALSE
     )
     return(NULL)
   }
-  if (is.null(eye)) {
-    eye <- getElem_eyecol(x)
+  if (is.null(eye_col)) {
+    eye_col <- getElem_eyecol(names(x))
   }
-
-  if (length(eye) > 1) {
-    warning("Please define eye column", call. = FALSE)
-    return(NULL)
-  }
-  if (length(eye) < 1) {
-    message("No eye column found: Counting patients only")
-    res <- c(patients = length(unique(x[[id]])))
-  } else if (length(eye) == 1) {
+  if (length(eye_col) != 1) {
+    message("Unclear which is the eye column. Counting id only.
+Specify eye column with \"eye_col\" argument")
+    res <- c(id = length(unique(x[[id_col]])))
+  } else if (length(eye_col) == 1) {
     quiet_recode <- purrr::quietly(~ recodeye(x = .x, dropunknown = .y))
-    recode_eye <- quiet_recode(.x = x[[eye]], .y = dropunknown)
+    recode_eye <- quiet_recode(.x = x[[eye_col]], .y = dropunknown)
     if (length(recode_eye$warnings) > 0) {
       message(recode_eye$warnings)
-      return(c(patients = length(unique(x[[id]]))))
+      return(c(id = length(unique(x[[id_col]]))))
     }
     if (length(recode_eye$messages) > 0) {
       message(gsub("\\\n", "", recode_eye$messages))
     }
-    if(any(unique(recode_eye$result) %in% "b")){
+    if(!details & any(unique(recode_eye$result) %in% "b")){
       message("Some rows contain information for both eyes. Correct?")
     }
-    x[[eye]] <- recode_eye$result
-    res <- count_eyes(x = x, id = id, eye = eye)
+    x[[eye_col]] <- recode_eye$result
+    res <- count_eyes(x = x, id_col = id_col, eye_col = eye_col, details = details)
   }
   res
 }
@@ -102,38 +107,53 @@ eyes <- function(x, id = NULL, eye = NULL, dropunknown = TRUE) {
 #' internal count
 #' @name count_eyes
 #' @param x object (data frame)
-#' @param id patient column
-#' @param eye eye column
-#' @description `count_eyes()` is the internal counting function
+#' @param id_col patient column
+#' @param eye_col eye column
+#' @param details if TRUE, will get more information
+#' @description `count_eyes()` is the internal counting function for [eyes]
 #' @return Named integer vector with count of patients and eyes
 #' @keywords internal
-#'   for [`eyes()`]
-count_eyes <- function(x, id, eye) {
-  n_pat <- length(unique(x[[id]]))
+count_eyes <- function(x, id_col, eye_col, details = FALSE) {
+  n_pat <- length(unique(x[[id_col]]))
 
   rl <- c("r", "l")
-  x[[eye]] <- factor(x[[eye]], levels = union(c("b", rl), unique(x[[eye]])))
-  eye_tab <- table(unique(x[, c(id, eye)]))
-
-  if (any(grepl("b", colnames(eye_tab)))) {
-    if (any(grepl(paste(rl, collapse = "|"), colnames(eye_tab)))) {
-      eye_tab[, rl][eye_tab[, "b"] == 1] <- 0
-    }
-  }
-
+  x[[eye_col]] <- factor(x[[eye_col]], levels = c("b", rl))
+  eye_tab <- table(unique(x[, c(id_col, eye_col)]))
+  eye_tab[, rl][eye_tab[, "b"] == 1] <- 0
   out <- colSums(eye_tab)
   outfinal <- out[rl] + out["b"]
   n_eyes <- sum(outfinal)
+if(!details){
   nr <- unname(outfinal["r"])
   nl <- unname(outfinal["l"])
-
-  return(c(patients = n_pat, eyes = n_eyes, right = nr, left = nl))
+  res <- list(id = n_pat, eyes = n_eyes, right = nr, left = nl)
+  class(res) <- c("eyes", class(res))
+  return(res)
+} else {
+  both <- eye_tab[, "b"] == 1 | rowSums(eye_tab) > 1
+  r_only <- eye_tab[, "r"] == 1 & rowSums(eye_tab) == 1
+  l_only <- eye_tab[, "l"] == 1 & rowSums(eye_tab) == 1
+  nb <- sum(both)
+  nr <- sum(r_only)
+  nl <- sum(l_only)
+  eyecount <- c(id = n_pat, eyes = n_eyes, right = nr, left = nl, both = nb)
+  patid <- rownames(eye_tab)
+  id <- list(right = patid[r_only], left = patid[l_only], both = patid[both])
+  res <- list(counts = as.list(eyecount), id = id)
+  class(res) <- c("eyes_details", class(res))
+  return(res)
+  }
 }
-
 #' @rdname eyes
 #' @inheritParams eyes
 #' @param ... passed to [eyes]
 #' @inheritParams eyes_to_string
+#' @section eyestr:
+#'  eyestr creates a string which can be pasted into reports.
+#'  It currently only supports "x eyes of n patient(s)" This is a limitation,
+#'   but I guess in the vast majority  of cases will be "correct".
+#'   To use for other categories (e.g., "people" or "participants"), use
+#'   `eyes(...)[1]`
 #' @return `eyestr`: Character string - can be directly pasted into reports
 #' @examples
 #' # Examples for the usage of eyestr
@@ -152,8 +172,8 @@ count_eyes <- function(x, id, eye) {
 #' lapply(ls_dat, eyestr, english = "none")
 #' @export
 eyestr <- function(x, ..., english = "small", caps = FALSE){
-  res <- suppressMessages(eyes(x, ...))
-    res_str <- res[c("patients", "eyes")]
+  res <- unlist(suppressMessages(eyes(x, ...)))
+    res_str <- res[c("id", "eyes")]
     eyes_to_string(res_str[!is.na(res_str)], caps = caps,
                    english = english)
   }