+++ R/print_methods.R
@@ -25,7 +25,7 @@ print.etdrs <- function(x, ...) {
 }
 
 #' @rdname print_methods
-#' @description S3 methods for blink class
+#' @description S3 methods for class blink
 #' @param x object of class "blink"
 #' @param ... arguments passed to [print.default]
 #' @export
@@ -60,3 +60,51 @@ print.blink <- function(x, ...) {
   cli::cli_h3("$IOP_eyes (right and left eyes)")
   print(x$IOP_eyes)
 }
+
+#' @rdname print_methods
+#' @description S3 methods for class eyes
+#' @param x object of class "eyes"
+#' @param ... arguments passed to [print.default]
+#' @export
+#' @importFrom cli rule
+print.eyes <- function(x, ...) {
+  cat(paste0(cli::rule(line = 2, left = "Counts", width = 25),
+             "\n"))
+  print(unlist(x))
+}
+#' @rdname print_methods
+#' @description S3 methods for class eyes_details
+#' @param x object of class "eyes_details"
+#' @param show how many subjects to be shown before printing the footnote
+#' @param ... arguments passed to [print.default]
+#' @export
+#' @importFrom cli rule
+#' @importFrom utils head
+print.eyes_details <- function(x, show = 6, ...) {
+  cat(paste0(cli::rule(line = 2, left = "$counts", width = 45),
+             "\n"))
+  print(unlist(x$counts))
+  cat(paste0("\n", cli::rule(line = 2, left = "$id", width = 45),
+             "\n"))
+  print_obj <- lapply(x$id, head)
+  ls_len <- lengths(x$id)
+  lapply(1:3, function(i){
+    if(ls_len[i] > show){
+      cat(paste0("$", names(print_obj)[i], "\n"))
+      print(print_obj[[i]])
+      newfoot <- paste0("with " , ls_len[i] - show, " more subjects\n\n")
+      cat(myfooter(newfoot))
+    } else{
+      print(print_obj[i])
+    }
+  })
+}
+
+#' @keywords internal
+#' @importFrom pillar style_subtle
+#' @importFrom cli symbol
+myfooter <- function(x) {
+  footer <- paste0(cli::symbol$ellipsis, " ", x)
+  pillar::style_subtle(paste("#", footer))
+}
+