+++ NEWS.md
@@ -1,4 +1,24 @@
-## Submission eye version 1.1.0  
+## 1.2.0
+Summary of changes: 
+- `eyes()` - now returns lists (of class "eyes") for easier access of count data
+- `set_eye_strings()` - set string codes globally! This makes it easier for people using different languages to use eye.
+- recodeye: renamed "eyecodes" argument to "eyestrings"
+- Change of terminology to "subjects/id" rather than "patients" 
+- simplified code
+
+Details: 
+- `eyes()`: "details" argument creates object of class "eyes_details", allowing access to in depth count statistics and subject ids 
+- This includes new print methods for class "eyes" and "eyes_details"
+- change id and eye arguments to "id_col" and "eye_col" - previous code should not break because of partial argument name matching. However, the new names make it clearer as to what those arguments are for.
+- calling `set_eye_strings` will super-assign the new codes to the internal
+list object eye_codes
+- getElem_eyecol now gives precedence to columns that are called "eye" or "eyes" (#30)
+Simplifications:
+- simplified pivoting in myop using the pivot_longer regex feature
+- removed ..._chr arguments from getElem_... functions (redundant)
+- getElem... now only for vectors (because they were anyways only used for vectors)
+
+## 1.1.0  
 summary of changes: 
 - new features for function `eyestr`
 - important bug fixes
@@ -13,10 +33,6 @@ Details:
 - added internal function `tidyNA_low` to avoid double tolower call
 - recodeye fixed fail when trailing white space (#28)
 - which_va fixed conversion error when mixed ETDRS and logmar (#29)
-<<<<<<< HEAD
-=======
-- update citation file
->>>>>>> cbdd7df1aa16373c3b0eeb89d62b39cf2b0d84f3
 
 # eye 1.0.1
 - Nasty bug fix! Erroneous conversion of quality visual acuity entries when ETDRS (#24)